Move vector-only components to src/components.
