+++ src/components/structures/RightPanel.js
